refactor: 사이드바 profile ui entities로 이동
