--- src/entities/user/api/index.ts
@@ -1,3 +1,0 @@
-// User API 세그먼트 Public API
-// export * from './user-api';
-// export * from './queries';
