--- src/entities/user/index.ts
@@ -1,4 +1,0 @@
-// User 엔티티 Public API
-export * from './model';
-// export * from './api';
-// export * from './ui';
