--- src/entities/user/model/__mocks__/user.mock.ts
@@ -1,8 +1,0 @@
-import { ProfileResponse } from '../types';
-
-export const mockProfile: ProfileResponse = {
-  id: '1234',
-  name: '정민희',
-  email: 'minhee@example.com',
-  image: null,
-};
