--- src/entities/user/model/index.ts
@@ -1,2 +1,0 @@
-export * from './queryKeys';
-export * from './types';
