--- src/entities/user/model/queryKeys.ts
@@ -1,5 +1,0 @@
-export const userQueryKeys = {
-  all: ['user'] as const,
-  profile: () => [...userQueryKeys.all, 'profile'] as const,
-  studies: () => [...userQueryKeys.all, 'studies'] as const,
-} as const;
