--- src/entities/user/ui/index.ts
@@ -1,4 +1,0 @@
-// User UI 세그먼트 Public API
-// export * from './UserCard';
-// export * from './UserProfile';
-// export * from './UserAvatar';
