--- src/features/get-profile/api/index.ts
@@ -1,1 +1,0 @@
-export * from './handler';
