+++ src/shared/mocks/handlers.ts
@@ -1,5 +1,4 @@
 import { noteHandlers } from '@/entities/note/model/mock/handler/noteHandler';
-import { sidebarProfileHandler } from '@/features/get-profile/api/handler';
 import { sidebarCreateStudyHandler } from '@/features/create-study/api/handler';
 import { sidebarStudyHandler } from '@/features/get-study-list/api/handler';
 import { sidebarGoalHandler } from '@/features/get-goal-list/api/handler';
@@ -15,7 +14,6 @@ import { deleteAccountHandler } from '@/features/delete-account/api/handler';
 import { logoutHandler } from '@/features/logout/api/handler';
 
 export const handlers = [
-  ...sidebarProfileHandler,
   ...sidebarCreateStudyHandler,
   ...sidebarStudyHandler,
   ...sidebarGoalHandler,
